refactor move MRIR to graphics pkg
